[None][fix] fused_cached_conv_ssm: remove stale int32 cast of slot_idx_dec
slot_idx is torch.long (int64). The .to(torch.int32) call created a copy
that was frozen in the CUDA graph memory pool at capture-time slot values.
During replay with different sequences at different cache slots, the Triton
kernel read stale slot indices => read/wrote SSM/conv state at wrong
positions => garbage decode outputs => GSM8k ~1.8%.

Removing the cast returns a view (int64 slice) that is updated in-place
by the resource handler before each CUDA graph replay. The Triton kernel
already does tl.load(...).to(tl.int64) internally so int64 input works
without any other change.

+++ tensorrt_llm/_torch/auto_deploy/custom_ops/mamba/fused_cached_conv_ssm.py
@@ -166,7 +166,7 @@ def _fused_cached_conv_ssm(
     if num_decode > 0:
         conv_in_dec = inp_flat[num_prefill_tokens:num_total_tokens]  # [nd, conv_dim]
         dt_dec = dt_flat[num_prefill_tokens:num_total_tokens]  # [nd, nheads]
-        slot_idx_dec = slot_idx[num_prefill:num_seq].to(torch.int32)
+        slot_idx_dec = slot_idx[num_prefill:num_seq]
 
         fused_conv_ssm_decode(
             conv_in_dec,
